+++ src/main/java/com/musala/drones/persistence/entity/dto/MedicationOutputDTO.java
@@ -2,13 +2,15 @@ package com.musala.drones.persistence.entity.dto;
 
 import java.util.UUID;
 
+import com.musala.drones.persistence.entity.enums.MedicationState;
+
 import lombok.Data;
 
 @Data
 public class MedicationOutputDTO {
     private UUID id;
     private String name;
-    private int weigth;
+    private double weigth;
     private String code;
     private MedicationState state;
 }
